refactor : MatchRepository 의 findByMatchId 파라미터 형 Long 수정에 따른 연쇄 변화 수정

+++ src/main/java/com/shootdoori/match/repository/MatchRepository.java
@@ -39,5 +39,5 @@ public interface MatchRepository extends JpaRepository<Match, Long> {
             Pageable pageable
     );
 
-  Match findByMatchId(int i);
+  Match findByMatchId(Long i);
 }
